Fix nil templates
There's a bug that errors out when a layout isn't available.

In the normal Rails rendering process, `nil` layouts are an acceptable value
and will render a template without a layout when that happens.

In form_props, we depend on
1. a layout yielding back a `json` object so that the template can continue to
render. With no layout, this suddenly errors out.
2. every template except for the layout to have a `json` local passed. We do this
by patching the `render` method on Rails, but this wasn't a great idea since it
does a double lookup and bled into non-json templates. It mayyy also conflict with
jbuilder which also does a `defined?` testing on `json`

To fix this, we change the rendering strategy. We still render top down, but instead
of testing for `defined?(json)` we use a instance variable called `@_props` in the
anonymous view and set a `json` local var to it. This has the immediate benefit of
no longer needing double template lookups, a less intrusive patch, and no longer
needing to assign locals which pollutes the dynamic method naming convention of template.

This resolves #1