Use a method instead of local_assigns (#10)
`virtual_path_of_template` was originally assigned to a local in hopes that the
layout would pick it up, but I didn't realize that the layout lookups passed an
empty `[]` as a locals so the template never picked it up. That's why I used
`local_assigns` which seemed like a hack.

Instead, we're going to def helper method on the anonymous view that's
generated with each request.