Remove auto camelize lower
This commit removes camelize lower on all keys. We had to remove this because
superglue deferrals were breaking on multiword keys like `your_trips`. This means
we have to be explicit about keys when working with templates.

Disregarding the bug, this is a good move as most people working with superglue
for the first time get confused about why the keys don't line up with what they
receive on the page component.

This also gives a nice speedup to props_template as we no longer have to deal with
camelize. Its now clearly the fastest json builder on rolftimmermans benchmark.

However, this is a backward breaking change. You must manually go through each key
and rename it to its `camelCase` equivalent. To restore the original behavior, use
something like the following:

```ruby
module PropsTemplateOverride
  def format_key(key)
    @key_cache ||= {}
    @key_cache[key] ||= key.camelize(:lower)
    @key_cache[key]
  end

  def result!
    result = super
    @key_cache = {}
    result
  end

  ::Props::BaseWithExtensions.prepend self
end
```