Remove virtual_path_of_template.
This method was originally created to more easily tie a page json with a
page component. It was to be used as the componentIdentifier.

A better and more maintainable approach would be to use a combination of
`controller_path` and `action` for the `componentIdentifier`. This
removes the hacky-ness of dynamically defining a method, while giving
flexibility as two controller actions can render the same template while
having a different componentIdentifier.