Active template virtual path (#13)
* Revert "Remove virtual_path_of_template. (#12)"

This reverts commit 55ace47646f2d3b4380180fcacbfe3a49ff37621.

* Expose method to all templates

After the revert, we'd like to rename the method and make it accessible to all
templates. The reasoning behind this is because superglue was using @virtual_path
on the html templates, which seemed like a private variable, with this change we
expose the active_template_virtual_path to everyone so we don't have to use private
instance vars elsewhere.