Fix Locals need to be shallow cloned. (#22)
Resolves #20. `prop_template` prebuilds an array of options before
handling it off to handle_collection to build. I recall this being
somewhat related to faster collection rendering. Unfortunately,
the trade off is that we have to be cognizant of how we mutate deeply
nested options as those can be the same object. In this case, I missed
a required clone on locals.