Add support for fragment context
This commit was missing from the 2.0 release and includes several related
improvements:

1. Adds fragmentContext to allow digging using props_at to find a node, then
considering what fragment it was found in, returns a `found_path` that can
be used in a graft response's `path`. We'll need to replace the generated
files in superglue_rails

2. Modifies `fragments!` to return paths that are relative to the data being
presented. So instead of `data.a.deep.path.in.page` it returns. `data.a.deep.path.in.response`
with the latter being much shorter.